Basic home page funcnality added

+++ app/src/main/java/com/github/danishjamal104/pizzaordy/data/model/CrustSize.kt
@@ -3,5 +3,5 @@ package com.github.danishjamal104.pizzaordy.data.model
 data class CrustSize(
     val id: Int,
     val name: String,
-    val price: Float
+    val price: Double
 )
